% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateNetwork.R
\name{annotate_go_bp}
\alias{annotate_go_bp}
\title{Add GO BP annotation to the graph vertices}
\usage{
annotate_go_bp(gg, annoF)
}
\arguments{
\item{gg}{graph to update}

\item{annoF}{annotation table}
}
\description{
Add GO BP annotation to the graph vertices
}
