% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateNetwork.R
\name{annotate_go_mf}
\alias{annotate_go_mf}
\title{Add GO MF annotation to the graph vertices}
\usage{
annotate_go_mf(gg, annoF)
}
\arguments{
\item{gg}{graph to update}

\item{annoF}{annotation table}
}
\description{
Add GO MF annotation to the graph vertices
}
