% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateNetwork.R
\name{annotate_SCHanno}
\alias{annotate_SCHanno}
\title{Title}
\usage{
annotate_SCHanno(gg, anno)
}
\arguments{
\item{anno}{}
}
\value{

}
\description{
Title
}
\examples{
file <- system.file("extdata", "PPI_Presynaptic.gml", package = "AnNet")
gg <- igraph::read.graph(file,format="gml")
afile<-system.file("extdata", "SCH_flatfile.csv", package = "AnNet")
dis    <- read.table(afile,sep="\t",skip=1,header=F,strip.white=T,quote="")
#agg<-annotate_SCHanno(gg,dis)
}
