% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateNetwork.R
\name{annotate_vertex}
\alias{annotate_vertex}
\title{Generic annotation function}
\usage{
annotate_vertex(gg, name, values)
}
\arguments{
\item{gg}{graph to annotate}

\item{name}{name of the attribute}

\item{values}{annotation data.frame}
}
\description{
It takes name of the attribute, and two column
data.frame with vertex ID in the first column and annotation in the second.
As a first step all attributes with provided names will be removed.
}
