% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateNetwork.R
\name{annotateGeneNames}
\alias{annotateGeneNames}
\title{Title}
\usage{
annotateGeneNames(gg)
}
\arguments{
\item{gg}{}
}
\description{
Title
}
\examples{
file <- system.file("extdata", "PPI_Presynaptic.gml", package = "AnNet")
gg <- igraph::read.graph(file,format="gml")
agg<-annotateGeneNames(gg)
}
