% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consMatrix.R
\name{buildConsensusMatFromFiles}
\alias{buildConsensusMatFromFiles}
\title{Build consensus matrix from set of files.}
\usage{
buildConsensusMatFromFiles(Dir, file.name, skip = 1, sep = "\\t")
}
\arguments{
\item{sep}{}
}
\description{
Build consensus matrix from set of files.
}
