% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildNetwork.R
\name{buildFromSynaptomeByEntrez}
\alias{buildFromSynaptomeByEntrez}
\title{Title}
\usage{
buildFromSynaptomeByEntrez(entrez)
}
\description{
Title
}
\examples{
cid<-match('Presynaptic',getCompartments()$Name)
t<-getAllGenes4Compartment(cid)
gg<-buildFromSynaptomeByEntrez(t$HumanEntrez)
}
