% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildNetwork.R
\name{buildFromSynaptomeGeneTable}
\alias{buildFromSynaptomeGeneTable}
\title{Title}
\usage{
buildFromSynaptomeGeneTable(t)
}
\arguments{
\item{t}{}
}
\description{
Title
}
\examples{
cid<-match('Presynaptic',getCompartments()$Name)
t<-getAllGenes4Compartment(cid)
gg<-buildFromSynaptomeByEntrez(t)
}
