% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{calcAllClustering}
\alias{calcAllClustering}
\title{Calculate memberships for all clustering algorithms and store them on the
graph vertices.}
\usage{
calcAllClustering(gg)
}
\arguments{
\item{gg}{}
}
\value{

}
\description{
Calculate memberships for all clustering algorithms and store them on the
graph vertices.
}
