% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCentrality.R
\name{calcCentrality}
\alias{calcCentrality}
\title{Calculate centrality measures for graph nodes and save them as vertex
property.}
\usage{
calcCentrality(gg)
}
\arguments{
\item{gg}{igraph object}
}
\value{
modified igraph object
}
\description{
Calculate centrality measures for graph nodes and save them as vertex
property.
}
\examples{
library(synaptome.db)
cid<-match('Presynaptic',getCompartments()$Name)
t<-getAllGenes4Compartment(cid)
gg<-buildFromSynaptomeByEntrez(t$HumanEntrez)
ggm<-calcCentrality(gg)
}
