% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCentrality.R
\name{calcCentralityExternalDistances}
\alias{calcCentralityExternalDistances}
\title{Function calculates matrix of distances between elements of list and
the reference matrix}
\usage{
calcCentralityExternalDistances(m, l, keepOrder = FALSE, dist = "euclidean")
}
\arguments{
\item{m}{reference matrix}

\item{l}{list of permuted matrix}

\item{keepOrder}{if FALSE valuess will be sorted}

\item{dist}{methods available from dist function}
}
\value{

}
\description{
Function calculates matrix of distances between elements of list and
the reference matrix
}
