% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCentrality.R
\name{calcCentralityInternalDistances}
\alias{calcCentralityInternalDistances}
\title{Function calculates matrix of distances between elements of list}
\usage{
calcCentralityInternalDistances(l, keepOrder = FALSE, dist = "euclidean")
}
\arguments{
\item{keepOrder}{if FALSE valuess will be sorted}

\item{dist}{methods available from dist function}
}
\description{
Function calculates matrix of distances between elements of list
}
