% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseasePairs.R
\name{calcDiseasePairs}
\alias{calcDiseasePairs}
\title{Title}
\usage{
calcDiseasePairs(
  gg,
  name,
  diseases = NULL,
  permute = c("none", "random", "binned")
)
}
\arguments{
\item{permute}{}
}
\description{
Title
}
