% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{calcMembership}
\alias{calcMembership}
\title{Calculate cluster memberships for the graph.}
\usage{
calcMembership(
  gg,
  alg = c("lec", "wt", "fc", "infomap", "louvain", "sgG1", "sgG2", "sgG5", "spectral")
)
}
\arguments{
\item{alg}{}
}
\description{
Calculate cluster memberships for the graph.
}
