% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{calcReclusterMatrix}
\alias{calcReclusterMatrix}
\title{Recluster graph}
\usage{
calcReclusterMatrix(gg, mem, alg, CnMAX)
}
\arguments{
\item{CnMAX}{}
}
\description{
Function takes graph \code{gg} and its membership matrix \code{mem}
as returned \code{calcMembership} and apply clustering algorithm \code{alg}
to all clusters larger than \code{CnMAX}
}
