% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildNetwork.R
\name{calcSparsness}
\alias{calcSparsness}
\title{Calculate sparsness of the graph.}
\usage{
calcSparsness(gg)
}
\arguments{
\item{gg}{graph to evaluate}
}
\value{
sparsness value
}
\description{
Calculate sparsness of the graph.
}
\examples{
cid<-match('Presynaptic',getCompartments()$Name)
t<-getAllGenes4Compartment(cid)
gg<-buildFromSynaptomeByEntrez(t)
calcSparsness(gg)
}
