% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateNetwork.R
\name{escapeAnnotation}
\alias{escapeAnnotation}
\title{Escapes elements of list in annotation, so they'll be searchable by grep.
In the case when annotation has not carefully planned, some annotations
could be substring of other, for example search fo DOID:14 could return
DOID:143, DOID:1433, and DOID:14330. To avoid this all names should be
enclosed in escape characters, which unlikely to find within annotation
itself.}
\usage{
escapeAnnotation(annVec, col = COLLAPSE, esc = ESC)
}
\arguments{
\item{annVec}{vector of annotation strings}

\item{col}{list separator character}

\item{esc}{escape character}
}
\description{
Escapes elements of list in annotation, so they'll be searchable by grep.
In the case when annotation has not carefully planned, some annotations
could be substring of other, for example search fo DOID:14 could return
DOID:143, DOID:1433, and DOID:14330. To avoid this all names should be
enclosed in escape characters, which unlikely to find within annotation
itself.
}
