% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitPowerLaw.R
\name{FitDegree}
\alias{FitDegree}
\title{Fit Power Law to degree distribution.}
\usage{
FitDegree(DEG, title, Nsim, DATAleg, WIDTH, HEIGHT)
}
\arguments{
\item{DEG}{degree distribution}

\item{HEIGHT}{}
}
\value{

}
\description{
Fit Power Law to degree distribution.
}
\examples{
##No: of bootstrap iterations
bootStrap    <- vector(length=3)
bootStrap[1] <- 100
bootStrap[2] <- 1000
bootStrap[3] <- 5000

b=2 #we'll stick with 1000 iteteractions

##Legend Titles
Legend <- vector(length=2)
Legend[1] <- "Presynaptic PPI"
Legend[2] <- "PSP PPI"

l=2;

#gg <- igraph::read.graph(sprintf("\%s/\%s/\%s.gml",OUT[3],subDIR[S],subDIR[S]),format="gml")
#pFit <- FitDegree( as.vector(igraph::degree(graph=gg)), subDIR[S], bootStrap[b], Legend[l], WIDTH, HEIGHT )
}
