% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridgeness.R
\name{getBridgeness}
\alias{getBridgeness}
\title{Calculate bridginess from consensus matrix}
\usage{
getBridgeness(gg, alg, conmat)
}
\arguments{
\item{gg}{igraph object}

\item{alg}{clustering algorithm}

\item{conmat}{consensus matrix calculated with that algorithm}
}
\value{

}
\description{
Calculate bridginess from consensus matrix
}
