% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{getClustering}
\alias{getClustering}
\title{Get clustering results for the graph.}
\usage{
getClustering(
  gg,
  alg = c("lec", "wt", "fc", "infomap", "louvain", "sgG1", "sgG2", "sgG5")
)
}
\arguments{
\item{alg}{}
}
\description{
Get clustering results for the graph.
}
