% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCentrality.R
\name{getCM}
\alias{getCM}
\title{Extracts particular measure from matrix and convert for distance calculation
by calcCentralityInternalDistances and calcCentralityExternalDistances
functions.}
\usage{
getCM(m, nm, keepOrder)
}
\arguments{
\item{m}{matrix of centrality measures as returned by getCentralityMatrix}

\item{nm}{name of the measure from m}

\item{keepOrder}{if FALSE valuess will be sorted}
}
\description{
Extracts particular measure from matrix and convert for distance calculation
by calcCentralityInternalDistances and calcCentralityExternalDistances
functions.
}
