% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturbationEntropy.R
\name{getEntropy}
\alias{getEntropy}
\title{Calculates perturbation entropy}
\usage{
getEntropy(gg, maxSr = NULL)
}
\arguments{
\item{gg}{igraph object}
}
\value{

}
\description{
Calculates perturbation entropy
}
\examples{
cid<-match('Presynaptic',getCompartments()$Name)
t<-getAllGenes4Compartment(cid)
gg<-buildFromSynaptomeByEntrez(t$HumanEntrez)
gg<-annotateGeneNames(gg)
e<- getEntropy(gg)
}
