% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCentrality.R
\name{getGraphCentralityECDF}
\alias{getGraphCentralityECDF}
\title{Convert centrality matrix into}
\usage{
getGraphCentralityECDF(m)
}
\arguments{
\item{m}{}
}
\description{
Convert centrality matrix into
}
