% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCentrality.R
\name{getRandomGraphCentrality}
\alias{getRandomGraphCentrality}
\title{Title}
\usage{
getRandomGraphCentrality(gg, type = c("gnp", "pa"), ...)
}
\arguments{
\item{type:}{\itemize{
\item gnp -- G(n,p) Erdos-Renyi model
\item pa --  Barabasi-Albert model
}}
}
\value{

}
\description{
Title
}
