% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memrob.R
\name{getRobustness}
\alias{getRobustness}
\title{Calculate cluster robustness from consensus matrix}
\usage{
getRobustness(gg, alg, conmat)
}
\arguments{
\item{gg}{igroph object}

\item{alg}{clustering algorithm}

\item{conmat}{consensus matrix}
}
\description{
Calculate cluster robustness from consensus matrix
}
