% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{layoutByRecluster}
\alias{layoutByRecluster}
\title{Calculate two-level layout from recluster matrix}
\usage{
layoutByRecluster(gg, remem, layout = layout_with_kk)
}
\arguments{
\item{layout}{}
}
\description{
Takes results of recluster and apply \code{layoutByCluster} to each
}
