% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateNetwork.R
\name{loopOverFiles}
\alias{loopOverFiles}
\title{Annotate graph from list of files}
\usage{
loopOverFiles(GG, FILES, NAME, IDS, addIDS)
}
\arguments{
\item{GG}{igraph object}

\item{FILES}{list of file path strings to read annotation from}

\item{NAME}{name of the vertex property}

\item{IDS}{vertex IDs}

\item{addIDS}{if TRUE NAME_ID property will be added}
}
\description{
Annotate graph from list of files
}
