% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consMatrix.R
\name{makeConsensusMatrix}
\alias{makeConsensusMatrix}
\title{Function to build consensus matrix in memory}
\usage{
makeConsensusMatrix(
  gg,
  N = 500,
  mask = 20,
  alg,
  type,
  reclust = FALSE,
  Cnmin = -1,
  Cnmax = 10
)
}
\arguments{
\item{gg}{graph to perturb}

\item{N}{number of perturbation steps}

\item{mask}{percentage of elements to perturbe}

\item{alg}{clustering alg.}

\item{type}{edges=>1 or nodes=>2  to mask}

\item{reclust}{logical to decide wether to invoke reclustering}

\item{Cnmin}{Cn min for Spectral algorithm}

\item{Cnmax}{Cn max for reclustering algorithms}
}
\value{
consensus matrix of Nvert X Nvert
}
\description{
Function to build consensus matrix in memory
}
\examples{
data(karate,package='igraphdata')
alg<-'louvain'
gg<-calcClustering(karate,alg = alg)
conmat<-makeConsensusMatrix(gg,N=100,mask = 10,alg = alg,type = 2)
dim(conmat)
}
