% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memrob.R
\name{memrob}
\alias{memrob}
\title{function to get the member robustness from the consensus matrices #BUG fixed 14/03/12 TIS}
\usage{
memrob(x, rm = data.frame())
}
\arguments{
\item{rm}{}
}
\value{

}
\description{
function to get the member robustness from the consensus matrices #BUG fixed 14/03/12 TIS
}
