% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normModularity.R
\name{normModularity}
\alias{normModularity}
\title{Calculates the normalized network modularity value.}
\usage{
normModularity(
  gg,
  alg = c("lec", "wt", "fc", "infomap", "louvain", "sgG1", "sgG2", "sgG5"),
  seed = NULL,
  Nint = 1000
)
}
\arguments{
\item{Nint}{}
}
\description{
Calculates the normalized network modularity value.
}
\examples{
cid<-match('Presynaptic',getCompartments()$Name)
t<-getAllGenes4Compartment(cid)
gg<-buildFromSynaptomeByEntrez(t$HumanEntrez)
nm<-normModularity(gg,alg='louvain')
}
