% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturbationEntropy.R
\name{plotEntropy}
\alias{plotEntropy}
\title{Plot entropy values}
\usage{
plotEntropy(SRprime, subTIT = "Entropy", SRo = NULL, maxSr = NULL)
}
\arguments{
\item{maxSr}{}
}
\description{
Plot entropy values
}
