% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseasePairs.R
\name{prepareGDA}
\alias{prepareGDA}
\title{Get particular annotation from the graph and format it to the suitable
form.}
\usage{
prepareGDA(gg, name)
}
\arguments{
\item{name}{}
}
\description{
Get particular annotation from the graph and format it to the suitable
form.
}
