% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseasePairs.R
\name{runPermDisease}
\alias{runPermDisease}
\title{Title}
\usage{
runPermDisease(
  gg,
  name,
  diseases = NULL,
  Nperm = 100,
  alpha = c(0.05, 0.01, 0.001)
)
}
\arguments{
\item{Nperm}{}
}
\description{
Title
}
