% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateNetwork.R
\name{unescapeAnnotation}
\alias{unescapeAnnotation}
\title{Perform opposite to escapeAnnotation operations: remove all escape
characters from annotation strings}
\usage{
unescapeAnnotation(annVec, col = COLLAPSE, esc = ESC)
}
\arguments{
\item{annVec}{vector of annotation strings}

\item{col}{list separator character}

\item{esc}{escape character}
}
\description{
Perform opposite to escapeAnnotation operations: remove all escape
characters from annotation strings
}
