% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateNetwork.R
\name{annotate_SCHanno}
\alias{annotate_SCHanno}
\title{Title}
\usage{
annotate_SCHanno(gg, anno)
}
\arguments{
\item{anno}{}
}
\description{
Title
}
\examples{
cid<-match('Presynaptic',getCompartments()$Name)
t<-getAllGenes4Compartment(cid)
gg<-buildFromSynaptomeByEntrez(t$HumanEntrez)
afile<-system.file("extdata", "SCH_flatfile.csv", package = "AnNet")
dis    <- read.table(afile,sep="\t",skip=1,header=FALSE,
strip.white=TRUE,quote="")
#agg<-annotate_SCHanno(gg,dis)
}
