% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateNetwork.R
\name{annotate_topOnto_ovg}
\alias{annotate_topOnto_ovg}
\title{Title}
\usage{
annotate_topOnto_ovg(gg, dis)
}
\arguments{
\item{dis}{}
}
\description{
Title
}
\examples{
cid<-match('Presynaptic',getCompartments()$Name)
t<-getAllGenes4Compartment(cid)
gg<-buildFromSynaptomeByEntrez(t$HumanEntrez)
afile<-system.file("extdata", "flatfile_human_gene2HDO.csv", package = "AnNet")
dis    <- read.table(afile,sep="\t",skip=1,header=F,strip.white=T,quote="")
#agg<-annotate_topOnto_ovg(gg,dis)
}
