% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{calcClustering}
\alias{calcClustering}
\title{Calculate memberships for particular clustering algorithms and store them on the
graph vertices.}
\usage{
calcClustering(gg, alg)
}
\arguments{
\item{alg}{algorithm to apply}
}
\description{
Calculate memberships for particular clustering algorithms and store them on the
graph vertices.
}
