% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitPowerLaw.R
\name{FitDegree}
\alias{FitDegree}
\title{Fit Power Law to degree distribution.}
\usage{
FitDegree(
  DEG,
  Nsim = 100,
  plot = FALSE,
  dir = ".",
  DATAleg = "Fit power-law",
  threads = 4,
  WIDTH = 480,
  HEIGHT = 480
)
}
\arguments{
\item{DEG}{degree distribution}

\item{threads}{}
}
\value{

}
\description{
Fit Power Law to degree distribution.
}
\examples{
##No: of bootstrap iterations
bootStrap <- 100

##Legend Titles
Legend <- "Presynaptic PPI"

dir<-'.'

file <- system.file("extdata", "PPI_Presynaptic.gml", package = "AnNet")
gg <- igraph::read.graph(file,format="gml")
pFit <- FitDegree( as.vector(igraph::degree(graph=gg)),threads=1)#, Nsim=bootStrap,dir=dir, DATAleg= Legend,plot=TRUE )
}
