% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCentrality.R
\name{getRandomGraphCentrality}
\alias{getRandomGraphCentrality}
\title{Title}
\usage{
getRandomGraphCentrality(gg, type = c("gnp", "pa", "cgnp", "rw"), ...)
}
\arguments{
\item{type:}{\itemize{
\item gnp -- G(n,p) Erdos-Renyi model
\item pa --  Barabasi-Albert model
\item cgnp -- new random graph from a given graph by randomly adding/removing edges
\item rw -- new random graph from a given graph by rewiring 25\% of edges preserving the degree distribution
}}
}
\description{
Title
}
\examples{
library(synaptome.db)
cid<-match('Presynaptic',getCompartments()$Name)
t<-getAllGenes4Compartment(cid)
gg<-buildFromSynaptomeByEntrez(t$HumanEntrez)
}
