% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{layoutByCluster}
\alias{layoutByCluster}
\title{Calculate layout based upon membership}
\usage{
layoutByCluster(gg, mem, layout = layout_with_kk)
}
\arguments{
\item{mem}{}
}
\description{
Calculate layout based upon membership
}
